HPB-4006 move to Validation namespace
